pyproject.toml
[tool.poetry]
name = "dj-social-auth-google"
version = "0.1.0"
description = ""
authors = ["r-lelis <lelis@labcodes.com.br>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
dj-rest-auth = {extras = ["with_social"], version = "^2.2.4"}
model-bakery = "^1.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
